feat: syscall: add new syscall `trace()`
This commit adds a new syscall named trace to the kernel. The
`trace()` syscall takes an integer mask argument which controls
the tarce mask. This mask determines which syscalls to trace. The
user can set it using the  numbers defined in `kernel/syscall.h`.

So for example to enable tracing `fork()` and `pipe()` calls:

`trace(1<<SYS_fork|1<<SYS_pipe);`

We have also added a `tracemask` field to the `proc` structure.
This stores the trace mask of the current process.

Signed-off-by: Jalaj <96870071+jalaj711@users.noreply.github.com>