feat: add `kpgtbl` syscall to print page table of a process
This commit adds the `kpgtbl()` syscall. This syscall allows
the user to print the page table to stdout. It can be handy for
debugging process.

Known Issue: According to this website:
https://pdos.csail.mit.edu/6.828/2024/labs/pgtbl.html

This syscall should give the output as

`.. .. ..0xfffffffffffff000: pte 0x000000002000184b pa 0x0000000080006000`

for the last page on the pagetable. However, the current output is

`.. .. .. 0x0000003ffffff000: pte 0x000000002000184b pa 0x0000000080006000`

The issue is in the virtual address part of the output. The leading
bits are all ones in the expected output but not in the actual output.

Signed-off-by: Jalaj <96870071+jalaj711@users.noreply.github.com>