feat: added `backtrace()` function
This function traces the complete sequence of function calls
that were made leading up to the current function. It prints
the return addresses of each of these functions. The output
from this function can be fed into the `addr2line` utility
or `riscv64-unknown-elf-addr2line` to trace the actual
function which called it.

We use the s0 register, which stores the current frame pointer,
to get this information. The `fp-8` lcoation on stack gives the
current return address and `fp-16` gives the previous frame
pointer.

Signed-off-by: Jalaj <96870071+jalaj711@users.noreply.github.com>