feat: syscall: added `sigalarm()` and `sigreturn()` syscalls
this commit adds the above two syscalls along with the userspace
utility `alarmtest` to test these functions. For context, the
signature for sigalarm is:

`sigalarm(int period, void (*handler)(void));`

and for sigreturn is:

`sigreturn(void);`

The basic idea is to call `handler` after every `period` timer
interrupts. The handler must call `sigreturn` to return to
original program counter from where it was called. If the
handler does not call sigreturn, control will not return to
original code and the kernel will not call the handler anymore,
even after `period` ticks have occurred.

Signed-off-by: Jalaj <96870071+jalaj711@users.noreply.github.com>