feat: make `fork()` calls copy-on-write
The calls to `fork()` are now made copy-on-write. We utilise the
RSW bit in the Sv39 PTE format of RISC-V to detect which pages
are using this feature. The kernel reserves some pages in the
beginning of the memory to keep reference count of each page.

The `uvmcopy` function now simply create a new pte in the new
process that maps to the same page rather than creating actual
physical copies of the pages. This saves time and memory on the
system. The actual copy is performed by the page fault trap
handler, which determines whether the page was a COW fault and
copies the page.

Two new functions are added in vm.c: `krefcnt()` and
`krefcntadd()`. The former returns the current ref count while
the latter increments the ref count of a given page.

Signed-off-by: Jalaj <96870071+jalaj711@users.noreply.github.com>